+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - The default profile no longer shows the amount of characters bound to it.
 - Cropped characteristic fields will now show a tooltip on hover with the full content of the field.
 - The description editor will now receive input focus after clicking any formatting tool button.
+- Customized names in chat frames will now be cropped at 35 characters.
 
 ## Fixed